used fast_t and least_t in Julian functions

+++ Julian/Julian.h
@@ -4,8 +4,8 @@
 #include "../common/CalendarCache.h"
 #include "../common/YMD.h"
 
-int JulianEncode(const struct CalendarCache* const restrict cache, struct YMD* const restrict output, const long udn);
+int JulianEncode(const struct CalendarCache* const restrict cache, struct YMD* const restrict output, const int_fast32_t udn);
 
-int JulianDecode(const struct CalendarCache* const restrict cache, long* output, const struct YMD* const restrict ymd);
+int JulianDecode(const struct CalendarCache* const restrict cache, int_fast32_t* output, const struct YMD* const restrict ymd);
 
 #endif
